+++ server.js
@@ -48,5 +48,3 @@ app.get("/", (req, res, next) => {
 })
 
 app.listen(5000);
-
-module.exports = app;
